backend/measurement/pyproject.toml
[project]
name = "novia-measurement"
version = "0.1.0"
description = "Body measurement service using HMR 2.0 + SMPL Anthropometry"
requires-python = ">=3.10,<3.12"
dependencies = [
    # PyTorch (CPU for Mac, MPS will be used automatically on Apple Silicon)
    "torch>=2.1.0",
    "torchvision>=0.16.0",

    # HMR 2.0 / 4D-Humans dependencies
    "numpy>=1.24.0,<2.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "scipy>=1.11.0",
    "scikit-image>=0.21.0",
    "trimesh>=4.0.0",

    # For ViT detection
    "timm>=0.9.0",
    "einops>=0.7.0",

    # Hugging Face for model downloads
    "huggingface-hub>=0.19.0",

    # Web service
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Utilities
    "httpx>=0.26.0",
    "tqdm>=4.66.0",
    "yacs>=0.1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/measurement"]

[tool.ruff]
line-length = 100
target-version = "py311"
